Rename heap.rs to allocator.rs
